Completely overhaul everything

+++ scripts/build_kernel.sh
@@ -3,31 +3,34 @@
 # exit on error
 set -e
 
-# include common functions
-# shellcheck disable=SC1091
-. "scripts/0_includes.sh"
+SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
 
 # include device-specific variables
-DEVICE="${1,,}"
-# shellcheck disable=SC1090
-. "devices/${DEVICE}.sh"
+DEVICE_ID="${1,,}"
 
-# import build markers
-BUILD_DATETIME="$(cat "data/${DEVICE}_build_datetime.txt")"
-BUILD_NUMBER="$(cat "data/${DEVICE}_build_number.txt")"
-GRAPHENE_RELEASE="$(cat "data/${DEVICE}_build_graphene.txt")"
-export BUILD_DATETIME BUILD_NUMBER
+echo "Checking if release needed..."
+python3 "$SCRIPT_DIR"/check_if_release_necessary.py "$BUILD_METADATA_FILE" && error_code=0 || error_code=$?
+case $error_code in
+    0) ;;
+    2) echo "Release already exists; release not needed."; exit 0;;
+    *) echo "Error when checking release (exit code ${error_code}), assuming release required. Continuing.";;
+esac
 
 ### BUILD KERNEL
 
 echo "=== Starting Kernel Build Process ==="
-echo "Device: ${DEVICE}"
-echo "GrapheneOS Release: ${GRAPHENE_RELEASE}"
+echo "Device: ${DEVICE_ID}"
+echo "GrapheneOS Release: ${GRAPHENEOS_VERSION}"
 
 # fetch kernel sources
 echo "Creating and entering kernel directory..."
-if [ ! -d "kernel" ]; then
-  git clone https://gitlab.com/grapheneos/kernel_pixel.git -b "refs/tags/${GRAPHENE_RELEASE}" --recurse-submodules kernel
+if [[ "$DEVICE_ID" == "dummy" ]]; then
+  echo "Dummy build, skipping initialization..."
+  mkdir -p kernel/
+elif [ -d "kernel" ]; then
+  echo "Kernel directory already exists. Skipping initialization..."
+else
+  eval "${KERNEL_FETCH_COMMAND}"
   pushd kernel/ || exit
 
     # remove abi_gki_protected_exports files
@@ -37,22 +40,26 @@ if [ ! -d "kernel" ]; then
     mkdir -p aosp
     # fetch & apply ksu and susfs patches
     pushd aosp/ || exit
-      echo "Fetching stock defconfig from ${KERNEL_IMAGE_REPO}..."
-      git clone --depth=1 --branch "${GRAPHENE_RELEASE}" --single-branch "${KERNEL_IMAGE_REPO}" kernel_image/
-      echo "Extracting kernel image configuration..."
-      lz4 -d kernel_image/grapheneos/Image.lz4 kernel_image/Image
-      ./scripts/extract-ikconfig kernel_image/Image > arch/arm64/configs/stock_defconfig
-      rm -rf kernel_image/
+      if [[ -v KERNEL_IMAGE_REPO ]]; then
+        echo "Fetching stock defconfig from ${KERNEL_IMAGE_REPO}..."
+        git clone --depth=1 --branch "${GRAPHENEOS_VERSION}" --single-branch "${KERNEL_IMAGE_REPO}" kernel_image/
+        echo "Extracting kernel image configuration..."
+        lz4 -d kernel_image/grapheneos/Image.lz4 kernel_image/Image
+        ./scripts/extract-ikconfig kernel_image/Image > arch/arm64/configs/stock_defconfig
+        rm -rf kernel_image/
+      else
+        echo "KERNEL_IMAGE_REPO not set, skipping fetching defconfig."
+      fi
 
       # apply kernelsu
       echo "Setting up KernelSU..."
-      curl -LSs "https://raw.githubusercontent.com/KernelSU-Next/KernelSU-Next/next/kernel/setup.sh" | bash -s ${KERNELSU_NEXT_BRANCH}
+      curl -LSs "https://raw.githubusercontent.com/KernelSU-Next/KernelSU-Next/next/kernel/setup.sh" | bash -s "${KERNELSU_BRANCH}"
 
       # hardcode kernelsu version
       pushd KernelSU-Next/ || exit
         echo "Configuring KernelSU version..."
-        KSU_VERSION=$(($(git rev-list --count HEAD) + 10200))
-        echo "Using KernelSU version: ${KSU_VERSION}"
+        KERNELSU_VERSION=$(($(git rev-list --count HEAD) + 10200))
+        echo "Using KernelSU version: ${KERNELSU_VERSION}"
         sed -i '/^ccflags-y += -DKSU_GIT_VERSION=/d' kernel/Makefile
         sed -i '1s/^/ccflags-y += -DKSU_GIT_VERSION='"${KSU_VERSION}"'\n/' kernel/Makefile
       popd || exit
@@ -82,7 +89,11 @@ if [ ! -d "kernel" ]; then
     patch -p1 < "../patches/0002-enable-wireguard-by-default.patch"
 
     echo "5. Applying 'stock defconfig spoof' patch..."
-    patch -p1 < "../patches/0003-spoof-stock-defconfig.patch"
+    if [[ -v KERNEL_IMAGE_REPO ]]; then
+      patch -p1 < "../patches/0003-spoof-stock-defconfig.patch"
+    else
+      echo "KERNEL_IMAGE_REPO not set, skipping stock defconfig spoof patch."
+    fi
 
     echo "6. Applying 'clean kernel version' patch..."
     patch -p1 < "../patches/0004-clean-kernel-version.patch"
@@ -95,15 +106,19 @@ fi
 pushd kernel/ || exit
   echo "=== Building Kernel ==="
   # build kernel
-  ${KERNEL_BUILD_COMMAND}
+  eval "${KERNEL_BUILD_COMMAND}"
 popd || exit
 
 echo "=== Moving Build Artifacts ==="
 # stash parts we need
 mv -v "kernel/out/${DEVICE_GROUP}/dist" "./kernel_out"
 
-echo "=== Cleaning Up ==="
-# remove kernel sources to save space before rom clone: SKIP FOR NOW
-#rm -rf kernel/
+# TODO: Can we get away with just the boot.img?
+
+echo "=== Zipping artifacts ==="
+pushd kernel_out/ || exit
+  mkdir -p ../build_output/
+  zip -r ../build_output/kernel-"${DEVICE_ID}"-"${GRAPHENEOS_VERSION}".zip .
+popd || exit
 
 echo "=== Kernel Build Complete ==="
